LINARO: Change debian/copyright to be GPLv2 only
The debian/copyright file in u-boot-linaro currently claims that u-boot
can be distributed under the terms of GPLv2 or any later version.
However, there are a number of files in this tree (including several
that originate in the Linux kernel) whose license is stated to be GPLv2
only, not GPLv2 or later.  Update the copyright file to state that the
effective license is v2, not "or later".

Signed-by: John Rigby <john.rigby@linaro.org>