ARMV7: Overo: Automatically set clock rate to maximum if mpurate env variable is "auto"
The maximum clock rate for the OMAP3 processors on Overo depends on the
processor type and revision.  This patch sets the clock rate to the
spec sheet maximum if the mpurate environment variable is set to
"auto".  Otherwise it passes the mpurate variable unchanged on the
kernel command line.

Signed-off-by: Steve Sakoman <steve.sakoman@linaro.org>
(cherry picked from commit 4c4ab9400afb80193968eeca89b5f49bbb9d2fbb)