ehci-hcd: Allow cleanups to happen on an EHCI
With this, the EHCI seems to "recover" from a timeout. This is
particularly
observable if you were to ping the wrong IP Address and then ping the
correct
one or if there was a temporary failure during tftp sessions.

All it takes is one timeout to disable it. If you have a noisy network
(lot
of traffic), even if the traffic is not for the board, the timeouts
don't occur.

Signed-off-by: Joel A Fernandes <agnel.joel@gmail.com>